refactor: outsource emoji data initialisation and settings store into common module
In order to be able to use it in the background script.
